renamed octree.hpp
