+++ domain/test/unit/tree/btree.cpp
@@ -32,7 +32,7 @@
 #include "gtest/gtest.h"
 
 #include "cstone/tree/btree.hpp"
-#include "cstone/tree/octree.hpp"
+#include "cstone/tree/csarray.hpp"
 #include "cstone/tree/octree_util.hpp"
 
 /*! @brief @file tests for binary tree generation
