+++ domain/test/unit/tree/octree_util.cpp
@@ -36,7 +36,7 @@
 #include "gtest/gtest.h"
 
 #include "cstone/sfc/morton.hpp"
-#include "cstone/tree/octree.hpp"
+#include "cstone/tree/csarray.hpp"
 #include "cstone/tree/octree_util.hpp"
 
 using namespace cstone;
