+++ domain/test/unit_cuda/octree.cu
@@ -36,7 +36,7 @@
 #include <thrust/device_vector.h>
 
 #include "coord_samples/random.hpp"
-#include "cstone/tree/octree.hpp"
+#include "cstone/tree/csarray.hpp"
 #include "cstone/tree/octree_gpu.h"
 #include "cstone/tree/octree_util.hpp"
 #include "cstone/tree/update_gpu.cuh"
